TST: Split sparse tests
